Initial attempt at supporting channels
Allows channels to be specified for each extension in a repository based on the package
ID.  Expects that pre-release versions are pushed with a --tag='channelName' to set
the dist-tag appropriately.  If not specified, the 'latest' channel is used by default.

Could also update the extension info view to indicate which channels are available for
tracking and couple this to a command that allows switching between channels w/o
manually editing settings.

#2