Finish release channels UI
Added a context menu entry for installed extensions to switch release
channels. This updates the privateExtensions.channels setting and asks
whether you want to update to the latest version on the new channel.

Replaced the channel label on the extension details page with a button
to switch channels. This only appears if there are multiple channels to
choose from.